Use shorts instead of char and int

+++ test/test_situation.cpp
@@ -28,7 +28,7 @@ TEST_CASE("Compute moves from a situation") {
 TEST_CASE("Compute parking from a situation") {
   Situation test("../test/support/puzzle.txt");
 
-  array<array<int, SIZE>, SIZE> expected_parking = {
+  array<array<short, SIZE>, SIZE> expected_parking = {
       {
           {
               1, -1, 2, 3, 3, 3
